# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nutrition Clinic API",
    "version": "2.0.0",
    "description": "واجهة REST لنظام إدارة عيادة التغذية. مصمَّمة ليستهلكها موقع الويب وتطبيق الموبايل (Android/iOS) الخاص بالمريض دون أي تعديل. المصادقة: POST /api/auth/login ثم إرسال `Authorization: Bearer <token>` في كل طلب. التوكن صالح 12 ساعة.",
    "contact": {
      "name": "Clinic IT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "تطوير محلي"
    },
    {
      "url": "https://api.clinic.example",
      "description": "الإنتاج"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/visits": {
      "get": {
        "tags": [
          "measurements"
        ],
        "summary": "زيارات مريض",
        "parameters": [
          {
            "name": "patient_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ items }"
          }
        }
      },
      "post": {
        "tags": [
          "measurements"
        ],
        "summary": "تسجيل زيارة",
        "responses": {
          "201": {
            "description": "Visit"
          }
        }
      }
    },
...
    "/api/payments": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "المدفوعات مع الإجمالي",
        "parameters": [
          {
            "name": "patient_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ items, total, currency }"
          }
        }
      },
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "تسجيل دفعة",
        "responses": {
          "201": {
            "description": "Payment"
          }
        }
      }
    },
    "/api/payments/summary": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "الإيرادات شهرياً/حسب الخدمة/حسب الطريقة",
        "parameters": [
          {
            "name": "months",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 12
            }
          }
        ],
        "responses": {
          "200": {
            "description": "RevenueSummary"
          }
        }
      }
    },
...
    "/api/reports/export/{kind}": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "تصدير CSV (patients|measurements|payments|appointments)",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "text/csv مع BOM يدعم العربية في Excel"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}